admin

+++ SpringBlog/Areas/Admin/Controllers/CategoriesController.cs
@@ -77,6 +77,12 @@ namespace SpringBlog.Areas.Admin.Controllers
                 return HttpNotFound();
             }
 
+            if (category.Posts.Count > 0)
+            {
+                TempData["ErrorMessage"] = "A category must not contain any posts in order to be deleted.";
+                return RedirectToAction("Index");
+            }
+
             db.Categories.Remove(category);
             db.SaveChanges();
             TempData["SuccessMessage"] = "The category has been deleted successfully.";
